# backend/Models/openAPI.yml
openapi: 3.0.3
info:
  title: Brewtiful API
  version: 1.0.0
servers:
  - url: http://{{baseurl}}
...
paths:
...
  /api/CartItem:
    get:
      summary: Get all cart items
      deprecated: false
      description: ''
      tags:
        - api/CartItem
      parameters:
        - name: Accept
          in: header
          description: ''
          required: true
          example: application/json
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            '*/*':
              schema:
                type: object
                properties:
                  id:
                    type: string
                  count:
                    type: string
                  itemId:
                    type: string
                  item:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      categoryId:
                        type: string
                      category:
                        type: object
                        properties:
                          id:
                            type: string
                          name:
                            type: string
                        required:
                          - id
                          - name
                      ingredientNames:
                        type: string
                      ingredients:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: string
                            name:
                              type: string
                            defaultQuantity:
                              type: string
                            extraCost:
                              type: string
                          required:
                            - id
                            - name
                            - defaultQuantity
                            - extraCost
                      price:
                        type: string
                    required:
                      - id
                      - name
                      - categoryId
                      - category
                      - ingredientNames
                      - ingredients
                      - price
                  cartId:
                    type: string
                  cart:
                    type: string
                  modifiedIngredients:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        quantity:
                          type: string
                  totalPrice:
                    type: string
                required:
                  - id
                  - count
                  - itemId
                  - item
                  - cartId
                  - cart
                  - modifiedIngredients
                  - totalPrice
              examples:
                '1':
                  summary: Get all cart items
                  value: "{\r\n    \"id\": \"integer\",\r\n    \"count\": \"integer\",\r\n    \"itemId\": \"integer\",\r\n    \"item\": {\r\n        \"id\": \"integer\",\r\n        \"name\": \"string\",\r\n        \"categoryId\": \"integer\",\r\n        \"category\": {\r\n            \"id\": \"integer\",\r\n            \"name\": \"string\"\r\n        },\r\n        \"ingredientNames\": \"array<string>\",\r\n        \"ingredients\": [\r\n            {\r\n                \"id\": \"integer\",\r\n                \"name\": \"string\",\r\n                \"defaultQuantity\": \"integer\",\r\n                \"extraCost\": \"float\"\r\n            },\r\n            {\r\n                \"id\": \"integer\",\r\n                \"name\": \"string\",\r\n                \"defaultQuantity\": \"integer\",\r\n                \"extraCost\": \"float\"\r\n            }\r\n        ],\r\n        \"price\": \"float\"\r\n    },\r\n    \"cartId\": \"integer\",\r\n    \"cart\": \"object\",\r\n    \"modifiedIngredients\": [\r\n        {\r\n            \"name\": \"string\",\r\n            \"quantity\": \"integer\"\r\n        }\r\n    ],\r\n    \"totalPrice\": \"float\"\r\n}\r\n"
          headers: {}
      security: []
    post:
      summary: Invalid fields (400)
      deprecated: false
      description: ''
      tags:
        - api/CartItem
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                Id:
                  type: integer
                Count:
                  type: integer
                ItemId:
                  type: integer
                Item:
                  type: object
                  properties:
                    Id:
                      type: integer
                    Name:
                      type: string
                    Price:
                      type: number
                    Category:
                      type: object
                      properties:
                        Id:
                          type: integer
                        Name:
                          type: string
                      required:
                        - Id
                        - Name
                    Ingredients:
                      type: array
                      items:
                        type: object
                        properties:
                          Id:
                            type: integer
                          Name:
                            type: string
                          DefaultQuantity:
                            type: integer
                          ExtraCost:
                            type: integer
                        required:
                          - Id
                          - Name
                          - DefaultQuantity
                          - ExtraCost
                    IngredientNames:
                      type: array
                      items:
                        type: string
                  required:
                    - Id
                    - Name
                    - Price
                    - Category
                    - Ingredients
                    - IngredientNames
                CartId:
                  type: integer
                Cart:
                  type: object
                  properties:
                    _id:
                      type: integer
                    Name:
                      type: string
                    IsSelected:
                      type: boolean
                    UserId:
                      type: integer
                    User:
                      type: object
                      properties:
                        Id:
                          type: integer
                        Name:
                          type: string
                        Points:
                          type: integer
                        Role:
                          type: string
                        Carts:
                          type: array
                          items:
                            type: string
                        UserLogin:
                          type: object
                          properties:
                            Token:
                              type: string
                            Password:
                              type: string
                            Username:
                              type: string
                          required:
                            - Token
                            - Password
                            - Username
                      required:
                        - Id
                        - Name
                        - Points
                        - Role
                        - Carts
                        - UserLogin
                    CartItems:
                      type: array
                      items:
                        type: string
                    Status:
                      type: string
                  required:
                    - _id
                    - Name
                    - IsSelected
                    - UserId
                    - User
                    - CartItems
                    - Status
                ModifiedIngredients:
                  type: array
                  items:
                    type: object
                    properties:
                      Name:
                        type: string
                      Quantity:
                        type: integer
                    required:
                      - Name
                      - Quantity
                TotalPrice:
                  type: number
              required:
                - Id
                - Count
                - ItemId
                - Item
                - CartId
                - Cart
                - ModifiedIngredients
                - TotalPrice
            example:
              Id: 12
              Count: 2
              ItemId: 8
              Item:
                Id: 1
                Name: Latte
                Price: 4.5
                Category:
                  Id: 1
                  Name: Kava
                Ingredients:
                  - Id: 1
                    Name: Pienas
                    DefaultQuantity: 1
                    ExtraCost: 0
                  - Id: 2
                    Name: Espresso
                    DefaultQuantity: 1
                    ExtraCost: 1
                IngredientNames:
                  - Pienas
                  - Espresso
              CartId: 1
              Cart:
                _id: 1
                Name: Domas123's Cart
                IsSelected: true
                UserId: 1
                User:
                  Id: 1
                  Name: Domas123
                  Points: 100
                  Role: Member
                  Carts: []
                  UserLogin:
                    Token: abcd1234
                    Password: password123
                    Username: Domas123
                CartItems: []
                Status: Active
              ModifiedIngredients:
                - Name: Espresso
                  Quantity: 4
                - Name: Pienas
                  Quantity: 1
              TotalPrice: 10.5
      responses:
        '400':
          description: ''
          content:
            '*/*':
              schema:
                type: object
                properties:
                  title:
                    type: string
                  status:
                    type: integer
                  detail:
                    type: string
                required:
                  - title
                  - status
                  - detail
              examples:
                '1':
                  summary: Invalid fields (400)
                  value: "{\r\n    \"title\": \"Bad Request\",\r\n    \"status\": 400,\r\n    \"detail\": \"Item not found.\"\r\n}"
          headers: {}
      security: []
...